requirement.txt
# --- Main App ---
streamlit
python-dotenv

# --- Pinecone ---
pinecone-client
langchain-pinecone

# --- Langchain & OpenAI (Modern, Compatible Versions) ---
langchain==0.3.27
langchain-core==0.3.79
langchain-openai==0.3.27
langchain-pinecone==0.2.13
langchain-text-splitters==0.3.11